Fix missing German translation for "Remove Vault" confirmation button
Fixes #3739

This pull request adds the missing German translation key `removeVault.confirmBtn` to the `strings_de.properties` file.

### Problem
In the German UI, the confirmation button when removing a vault was displayed in English ("Remove Vault") instead of the expected German translation.

### Cause
The key `removeVault.confirmBtn` was present in the default (English) translation file, but missing from the German one. When a key is not found in the localized file, Cryptomator falls back to the default – hence the English text appeared.

### Solution
Added the following line to `strings_de.properties`:

```strings_de.properties
removeVault.confirmBtn=Tresor entfernen
```

Notes
This is my first open source contribution – feedback is very welcome!


Fixes #3739